requirements.txt
pip==22.2.1
mysql-connector-python~=8.0.30
python-dotenv==0.20.0
requests~=2.28.1
pika~=1.3.0